+++ src/main/java/com/infoshareacademy/drinkers/App.java
@@ -3,6 +3,8 @@ package com.infoshareacademy.drinkers;
 import com.infoshareacademy.drinkers.domain.drink.Drink;
 import com.infoshareacademy.drinkers.service.filtering.FilterElements;
 import com.infoshareacademy.drinkers.service.filtering.FilterList;
+import com.infoshareacademy.drinkers.service.gson.JsonReader;
+import com.infoshareacademy.drinkers.service.printing.PrintElement;
 import com.infoshareacademy.drinkers.service.properties.AppProperties;
 import com.infoshareacademy.drinkers.service.properties.ApplicationProperties;
 import com.infoshareacademy.drinkers.service.sorting.SortDrinks;
