+++ src/main/java/com/infoshareacademy/drinkers/service/gson/GsonCreator.java
@@ -1,4 +1,4 @@
-package com.infoshareacademy.drinkers;
+package com.infoshareacademy.drinkers.service.gson;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
