+++ src/main/java/com/infoshareacademy/drinkers/service/gson/JsonReader.java
@@ -1,4 +1,4 @@
-package com.infoshareacademy.drinkers;
+package com.infoshareacademy.drinkers.service.gson;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonSyntaxException;
